Override browser fetch to use in tracking automatically
