+++ frontend/src/routes/events/[slug]/leaderboard/+page.ts
@@ -2,10 +2,9 @@ import { error } from "@sveltejs/kit";
 import type { PageLoad } from "./$types";
 import { client, EventsService } from "$lib/client/sdk.gen";
 import type { Project } from "$lib/client/types.gen";
-import { createTrackingFetch } from "$lib/tracking-fetch";
 
 export const load: PageLoad = async ({ params, fetch, parent }) => {
-  client.setConfig({ fetch: createTrackingFetch(fetch) });
+  client.setConfig({ fetch });
 
   const { event } = await parent();
   const {
